+++ database/migrations/2025_10_14_105148_create_likes_table.php
@@ -4,8 +4,7 @@ use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
@@ -15,6 +14,8 @@ return new class extends Migration
             $table->id();
             $table->foreignId('blueprint_id')->constrained()->cascadeOnDelete()->references('id')->on('blueprints');
             $table->foreignId('user_id')->constrained()->cascadeOnDelete()->references('id')->on('users');
+            $table->unique(['blueprint_id', 'user_id']);
+            $table->timestamps();
         });
     }